pyproject.toml
[project]
name = "customer-complaint-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "deprecated==1.3.1",
    "flask==3.1.2",
    "flask-cors==6.0.2",
    "flask-limiter==4.1.1",
    "flask-login==0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf==1.2.2",
    "greenlet==3.3.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "limits==5.6.0",
    "markupsafe==3.0.3",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "pdfkit==1.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "sqlalchemy==2.0.45",
    "typing-extensions==4.15.0",
    "werkzeug==3.1.4",
    "wrapt==2.0.1",
    "wtforms==3.2.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.8.0",
]
